Implement initial support for iterable declarations
That is, adds support for iterable<K, V> syntax, by delegating to the [Symbol.iterator]() method on the impl class, which is assumed to return something of the correct form.

This will generate the [Symbol.iterator](), keys(), values(), entries(), and forEach() methods on the wrapper class.

This initial implementation is somewhat inefficient (e.g. each call to next() and each iteration of forEach() does a O(n) iteration-and-copy over the original impl's list of keys and values) but we can change that later.

Also updates webidl-conversions to ^3.0.0.

Fixes #30.