Add support for running the test command in watch mode
When running the test in watch mode, via yarn test --watch, jest enters in an infinite loop. This is due to the fact that the test/output folder is watched by jest. Exclude this folder from the watched paths.

Also, .gitignore the coverage folder generated when running yarn test --coverage.