Make it possible to have depency cycles involving the generated files
This commit switches the generated files from using module.exports to using exports. This means that, as long as imports are referenced in certain careful ways, such as

const Foo = require("./generated/Foo");

// later
Foo.isImpl()

instead of

const isFooImpl = require("./generated/Foo").isImpl;

// later
isFooImpl()

that the generated files can now be used as part of dependency cycles. Although this is fragile and not a pattern that should be used often, it can be important for implementing certain parts of the web platform which involve circular dependencies at the spec level.