Add proper reflection support through processor (#161)
It has long been proposed that reflection features be moved to the
host-side, considering it is a part of HTML rather than Web IDL. With
the recently added processor infrastructure, this is now possible.

With this change, non-relative imports are now allowed in processors'
addImport function. This is done by checking if the imported package
fits the npm package name regular expression. Thanks to ExE Boss for
this contribution.

We also note that existing reflectors did not deal very well with the
change from this to esValue, since
2a2a0999a950106094e83ec9d12ea91bad51b0e2. Since this commit moves the
reflection code out of webidl2js, the bug now naturally disappears.

Closes #15. Closes #93.

Co-authored-by: ExE Boss <3889017+ExE-Boss@users.noreply.github.com>
Co-authored-by: Domenic Denicola <d@domenic.me>