Tweak test cases for dictionary defaults
In Web IDL these days, optional dictionary arguments must have a trailing "= {}". This adds those to the test .webidl files. Doing so uncovered an issue where DOMRectInit was referenced in the test files, but did not exist, so we add code to throw in this case, and correct the test file.